Factor out SSL context creation, improve SSL log message
